Day 7
It’s the weekend, but I really wanted to make more progress on my code. I worked out the character movement and blitting for player one, but I still have not figured out why player two is not working. But, I also found images for the health bars and blitted the starting ones.